% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gwcolumn}
\alias{gwcolumn}
\title{Create a gwcolumn within a UI definition}
\usage{
gwcolumn(nominator = "", denominator = c(1:12), ..., class = NULL)
}
\arguments{
\item{numerator}{}

\item{denominator}{}

\item{...}{Elements to include within the column}

\item{class}{additional class definitions to add.}
}
\value{
A gwcolumn that can be included within a
\code{\link{gwfluidRow}}.
}
\description{
Create a gwcolumn for use within a \code{\link{gwfluidRow}}
}
\seealso{
\code{\link{gwfluidRow}}, \code{\link{gwfixedRow}}.
}

