% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gwfluidPage}
\alias{gwfluidPage}
\alias{gwfluidRow}
\title{Create a page to utilise Groundworks grids}
\usage{
gwfluidPage(...)

gwfluidRow(..., class = NULL)
}
\arguments{
\item{...}{Elements to include within the page}

\item{class}{additional class definitions to add.}
}
\description{
Add detail here
}

