% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gwLIBS}
\alias{gwLIBS}
\title{Add Groundwork libraries}
\usage{
gwLIBS(...)
}
\arguments{
\item{...}{optional additional tags}
}
\description{
\code{gwLIBS}
A utility function to add the necessary CSS and js libraries for the GroundWorks CSS framework.
}
\examples{
\dontrun{
gwLIBS()
}
}

