% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gwMenuBar}
\alias{gwMenuBar}
\title{Create a GroundWorks navigation menu bar.}
\usage{
gwMenuBar(title, ..., id = NULL)
}
\arguments{
\item{title}{Title of the navbar menu.}

\item{...}{Elements to include in the navbar menu.}

\item{id}{The unique id for this element in the DOM.}

\item{gridBreakpoint}{By default, grid columns collapse into rows at the $small-tablet breakpoint. The grid breakpoint helpers allow you to target the breakpoint that the grid structure should persist through.
For more detail see \link{http://groundworkcss.github.io/groundwork/docs/breakpoints.htm}}
}
\description{
Create a GroundWorks navigation menu bar. for use within a \code{\link{gwfluidPage}}
}

