% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gwNavButton}
\alias{gwNavButton}
\title{Create a GroundWorks navigation button.}
\usage{
gwNavButton(title, id = NULL, class = NULL)
}
\arguments{
\item{title}{Title of the navbar button.}

\item{id}{The unique id for this element in the DOM.}
}
\description{
Create a GroundWorks navigation button. For use within a \code{\link{gwfluidPage}}
}

