% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gwfluidPage}
\alias{gwfluidPage}
\alias{gwfluidRow}
\title{Create a page to utilise Groundworks grids}
\usage{
gwfluidPage(...)

gwfluidRow(...)
}
\arguments{
\item{...}{Elements to include within the page}
}
\description{
Add detail here
}

